requirements.txt
PyOpenGL
PyOpenGL-Accelerate
glfw